Sequencer Scripting: Change to get/set view and working ranges to double
#jira UE-192832
#rb david.bromberg

[CL 27240177 by max chen in ue5-main branch]

+++ Engine/Plugins/Animation/AnimationData/Source/AnimationData/Private/AnimSequencerController.cpp
@@ -230,10 +230,10 @@ void UAnimSequencerController::SetMovieSceneRange(FFrameNumber InFrameNumber) co
 	MovieScene->SetPlaybackRangeLocked(false);
 	
 	FMovieSceneEditorData& EditorData = MovieScene->GetEditorData();
-	EditorData.ViewStart = EditorData.WorkStart = 0.f;
+	EditorData.ViewStart = EditorData.WorkStart = 0.0;
 	EditorData.ViewEnd = EditorData.WorkEnd = Model->GetFrameRate().AsSeconds(InFrameNumber);
-	MovieScene->SetWorkingRange(static_cast<float>(EditorData.ViewStart), static_cast<float>(EditorData.ViewEnd));
-	MovieScene->SetViewRange(static_cast<float>(EditorData.ViewStart), static_cast<float>(EditorData.ViewEnd));
+	MovieScene->SetWorkingRange(EditorData.ViewStart, EditorData.ViewEnd);
+	MovieScene->SetViewRange(EditorData.ViewStart, EditorData.ViewEnd);
 
 	if (UMovieSceneControlRigParameterSection* Section = Model->GetFKControlRigSection())
 	{
